+++ migrator/update-application-js-step.spec.js
@@ -1,0 +1,14 @@
+/* eslint-env jest */
+
+describe('update application js steps', () => {
+  beforeEach(() => {
+  })
+
+  afterEach(() => {
+    jest.clearAllMocks()
+  })
+  
+  it('', () => {
+
+  })
+})